requirements.txt
coverage < 4.0.0
Django == 1.11b1
djangorestframework == 3.5.3
mock < 2.0.0
nose < 2.0.0
python-coveralls < 3.0.0
unittest2 < 1.0.0
